requirements.txt
# Core ML libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scikit-misc>=0.1.4
scipy>=1.7.0

# Deep learning
torch>=1.11.0
torch-geometric>=2.0.0
pytorch-lightning>=1.9.0
transformers>=4.20.0

# Single-cell analysis
scanpy>=1.8.0
anndata>=0.8.0
h5py>=3.6.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.62.0
joblib>=1.1.0
requests>=2.28.0

# Jupyter
# jupyter>=1.0.0  # Commented out since PyCharm edition doesn't support notebooks
# ipywidgets>=7.6.0  # Not needed without Jupyter

# Logging and experiment tracking
wandb>=0.13.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0

# Development
black>=22.0.0
flake8>=4.0.0

# Additional dependencies for your existing scripts
gseapy>=1.0.0  # For GO analysis in Random Forest script

# Additional dependencies for STATE model implementation
einops>=0.6.0  # For tensor operations
torchmetrics>=0.11.0  # For metrics
hydra-core>=1.3.0  # For configuration management
omegaconf>=2.3.0  # For configuration
lightning>=2.0.0  # Updated lightning

# Note: pickle, gc, os, sys, json, warnings are built-in Python modules